move DataType to own class & getElementType util to JsonUtils
