+++ lib/src/main/java/growthbook/sdk/java/services/ConditionEvaluator.java
@@ -5,6 +5,7 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
 import com.google.gson.reflect.TypeToken;
+import growthbook.sdk.java.models.DataType;
 import growthbook.sdk.java.models.Operator;
 import growthbook.sdk.java.models.UserAttributes;
 
@@ -15,28 +16,6 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class ConditionEvaluator implements IConditionEvaluator {
-
-    enum DataType {
-        STRING("string"),
-        NUMBER("number"),
-        BOOLEAN("boolean"),
-        ARRAY("array"),
-        OBJECT("object"),
-        NULL("null"),
-        UNDEFINED("undefined"),
-        UNKNOWN("unknown"),
-        ;
-        private final String rawValue;
-
-        DataType(String rawValue) {
-            this.rawValue = rawValue;
-        }
-
-        @Override
-        public String toString() {
-            return this.rawValue;
-        }
-    }
 
     private final GrowthBookJsonUtils jsonUtils = GrowthBookJsonUtils.getInstance();
 
@@ -158,32 +137,6 @@ public class ConditionEvaluator implements IConditionEvaluator {
         return element;
     }
 
-    DataType getType(@Nullable JsonElement element) {
-        try {
-            if (element == null) return DataType.UNDEFINED;
-
-            if (element.toString().equals("null")) {
-                return DataType.NULL;
-            }
-
-            if (element.isJsonPrimitive()) {
-                JsonPrimitive primitive = element.getAsJsonPrimitive();
-                if (primitive.isBoolean()) return DataType.BOOLEAN;
-                if (primitive.isNumber()) return DataType.NUMBER;
-                if (primitive.isString()) return DataType.STRING;
-            }
-
-            if (element.isJsonArray()) return DataType.ARRAY;
-            if (element.isJsonObject()) return DataType.OBJECT;
-
-            return DataType.UNKNOWN;
-        } catch (RuntimeException e) {
-            e.printStackTrace();
-            return DataType.UNKNOWN;
-        }
-    }
-
-
 
     /**
      * Evaluates the condition using the operator. For example, if you provide the following condition:
@@ -241,7 +194,7 @@ public class ConditionEvaluator implements IConditionEvaluator {
         Operator operator = Operator.fromString(operatorString);
         if (operator == null) return false;
 
-        DataType attributeDataType = getType(actual);
+        DataType attributeDataType = GrowthBookJsonUtils.getElementType(actual);
 
         switch (operator) {
             case IN:
@@ -377,7 +330,7 @@ public class ConditionEvaluator implements IConditionEvaluator {
                 return !evalConditionValue(expected, actual);
 
             case TYPE:
-                return getType(actual).toString().equals(expected.getAsString());
+                return GrowthBookJsonUtils.getElementType(actual).toString().equals(expected.getAsString());
 
             case EXISTS:
                 boolean exists = expected.getAsBoolean();
